Development (#2)
* Computer class

This is mostly related to the computer class. I have verified that the reboot required part works correctly. The checking for updates works per say but it is so unreadable. I will need to find a way to parse that information.

* added to git ignore

I have added the __pycache__ to the .gitignore just in case it causes any issues in the future. I would rather this be created each time when they run the application for the first time.